portfolio_backend/requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
charset-normalizer==3.0.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.5
django-cuser==2017.3.16
django-debug-toolbar==3.2.2
django-filter==2.4.0
django-rest-swagger==2.1.2
djangorestframework==3.14.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
openapi-codec==1.3.2
psycopg2-binary==2.9.1
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
simplejson==3.18.1
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.14
